requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
accelerate==1.2.1
    # via project-coding (pyproject.toml)
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.2
    # via aiohttp
attrs==24.3.0
    # via aiohttp
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==3.2.0
    # via project-coding (pyproject.toml)
diffusers==0.32.1
    # via project-coding (pyproject.toml)
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
filelock==3.16.1
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fonttools==4.55.3
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
ftfy==6.3.1
    # via project-coding (pyproject.toml)
huggingface-hub==0.27.0
    # via
    #   accelerate
    #   datasets
    #   diffusers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via diffusers
jinja2==3.1.5
    # via torch
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via project-coding (pyproject.toml)
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.4.2
    # via torch
numpy==2.2.1
    # via
    #   project-coding (pyproject.toml)
    #   accelerate
    #   contourpy
    #   datasets
    #   diffusers
    #   matplotlib
    #   pandas
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   transformers
pandas==2.2.3
    # via datasets
pillow==11.0.0
    # via
    #   diffusers
    #   matplotlib
    #   torchvision
pip==24.3.1
    # via project-coding (pyproject.toml)
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psutil==6.1.1
    # via accelerate
pyarrow==18.1.0
    # via
    #   project-coding (pyproject.toml)
    #   datasets
pyparsing==3.2.0
    # via matplotlib
pyqt5==5.15.11
    # via project-coding (pyproject.toml)
pyqt5-qt5==5.15.16
    # via pyqt5
pyqt5-sip==12.16.1
    # via pyqt5
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   transformers
safetensors==0.4.5
    # via
    #   accelerate
    #   diffusers
    #   transformers
setuptools==75.6.0
    # via torch
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
tokenizers==0.21.0
    # via transformers
torch==2.5.1
    # via
    #   project-coding (pyproject.toml)
    #   accelerate
    #   torchvision
torchvision==0.20.1
    # via project-coding (pyproject.toml)
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.47.1
    # via project-coding (pyproject.toml)
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via ftfy
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
